requirements.txt
Flask
Flask-Session
python-dotenv
langchain
Flask-SocketIO